chore: advanced中type的int对应的int型调整为long型

+++ MFAAvalonia/Extensions/MaaFW/MaaInterfaceAdvancedOption.cs
@@ -31,7 +31,7 @@ public class MaaInterfaceAdvancedOption
             var typeName = type.ToLower();
             typeMap[Field[i]] = typeName switch
             {
-                "int" => typeof(int),
+                "int" => typeof(long),
                 "float" => typeof(float),
                 "bool" => typeof(bool),
                 _ => typeof(string)
